requirements.txt
pandas==2.2.3
streamlit==1.39.0
plotly==5.24.1
python-dateutil==2.9.0.post0